+++ fe/fe-core/src/test/java/com/starrocks/external/iceberg/IcebergHiveCatalogTest.java
@@ -13,6 +13,7 @@ import org.apache.iceberg.Table;
 import org.apache.iceberg.catalog.Catalog;
 import org.apache.iceberg.catalog.TableIdentifier;
 import org.apache.iceberg.hive.HiveCatalog;
+import org.apache.iceberg.relocated.com.google.common.collect.Maps;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -52,4 +53,13 @@ public class IcebergHiveCatalogTest {
         Assert.assertTrue(table.name().equals("test"));
     }
 
+    @Test
+    public void testInitialize() {
+        try {
+            IcebergHiveCatalog catalog = new IcebergHiveCatalog();
+            catalog.initialize("hive", Maps.newHashMap());
+        } catch (Exception e) {
+            Assert.fail(e.getMessage());
+        }
+    }
 }
